requirements.txt
# Core dependencies
numpy>=1.19.0
scipy>=1.6.0
pillow>=8.0.0
opencv-python>=4.5.0
python-magic>=0.4.0

# Utility dependencies
requests>=2.25.0

# ML dependencies - requires Python 3.10-3.12
tensorflow>=2.13.0
matplotlib>=3.4.0

# Optional dependencies for enhanced functionality
scikit-learn>=1.0.0  # For additional ML capabilities
tqdm>=4.60.0         # For progress bars during model downloads and processing

# IMPORTANT: TensorFlow 2.13.0+ requires Python 3.10-3.12
# If you're using Python 3.13+, please downgrade your Python version:
# 1. Create a new virtual environment: python3.10 -m venv steg-env-py310
# 2. Activate it: source steg-env-py310/bin/activate (Linux/macOS) or .\steg-env-py310\Scripts\activate (Windows)
# 3. Install dependencies: pip install -r requirements.txt
